requirements.txt
alembic==1.9.0
anyio==3.6.2
asyncpg==0.27.0
attrs==22.2.0
certifi==2022.12.7
cfgv==3.3.1
click==8.1.3
distlib==0.3.6
exceptiongroup==1.0.4
factory-boy==3.2.1
Faker==15.3.4
fastapi==0.88.0
filelock==3.8.2
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.1
identify==2.5.11
idna==3.4
inflection==0.5.1
iniconfig==1.1.1
Mako==1.2.4
MarkupSafe==2.1.1
nodeenv==1.7.0
packaging==22.0
platformdirs==2.6.0
pluggy==1.0.0
pre-commit==2.20.0
psycopg2-binary==2.9.5
pydantic==1.10.2
pytest==7.2.0
pytest-asyncio==0.20.3
python-dateutil==2.8.2
python-dotenv==0.21.0
PyYAML==6.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a30
sqlmodel==0.0.8
starlette==0.22.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.20.0
virtualenv==20.17.1
